色々修正

+++ src/misc/common/irange.hpp
@@ -5,7 +5,7 @@ class irange
 private:
     struct itr
     {
-        itr(int start = 0, int step = 1) : m_cnt{start}, m_step{step} {}
+        itr(i64 start = 0, i64 step = 1) : m_cnt{start}, m_step{step} {}
         bool operator!=(const itr& it) const
         {
             return m_cnt != it.m_cnt;
@@ -19,17 +19,17 @@ private:
             m_cnt += m_step;
             return *this;
         }
-        int m_cnt, m_step;
+        i64 m_cnt, m_step;
     };
-    int m_start, m_end, m_step;
+    i64 m_start, m_end, m_step;
 
 public:
-    irange(int start, int end, int step = 1)
+    irange(i64 start, i64 end, i64 step = 1)
     {
         assert(step != 0);
-        const int d = std::abs(step);
-        const int l = (step > 0 ? start : end);
-        const int r = (step > 0 ? end : start);
+        const i64 d = std::abs(step);
+        const i64 l = (step > 0 ? start : end);
+        const i64 r = (step > 0 ? end : start);
         int n = (r - l) / d + ((r - l) % d ? 1 : 0);
         if (l >= r) { n = 0; }
         m_start = start;
